Added Server
Added the server to communicated with Google API. React Native currently does not have anything other than a fetch() method, therefore oauth2.0 is only done by 3rd party services. To get around this, there will be a Node JS server running that the fetch command will communicate with to grab the data. To secure this, a different method will be used. In it's current state, it is insecure.